Bug fix: seq duration was passed incorrectly, so it ended one iter earlier than parallel in some cases

+++ seq.cpp
@@ -158,7 +158,7 @@ void visualize(std::vector<Body> &bodies) {
 }
 
 
-bool collisionTest(std::vector<Body> &bodies, int duration) 
+bool collisionTest(std::vector<Body> &bodies, float duration) 
 {
     bool collisionDetected = false;
     int timestepCounter = 0;
